DYN-9707: Rename graph-lock result types and clarify conflict message (M2, M8)
- M2: split GraphLockResults.cs into one type per file (GraphLockOutcome.cs,
  GraphLockAcquireResult.cs, GraphLockUserResponse.cs); rename the enum
  GraphLockResults -> GraphLockOutcome (singular, .NET convention) and the
  GraphLockAcquireResult.Conflict property -> Outcome, since values like
  Acquired are not conflicts.
- M8: reword GraphLockFileAlreadyOpenMessage so it no longer presumes a
  workspace exists or tells the user to close the other instance, and remove
  the leading whitespace before the bullet characters.
