+++ src/DynamoCore/Graph/Workspaces/Locking/GraphLockOutcome.cs
@@ -1,0 +1,12 @@
+namespace Dynamo.Graph.Workspaces.Locking
+{
+    /// <summary>
+    /// Describes the outcome of trying to acquire a graph lock.
+    /// </summary>
+    internal enum GraphLockOutcome
+    {
+        Acquired,
+        Cancelled,
+        Unavailable
+    }
+}
